More sync logic
